Check if urls are already archived w/ IA API
Often, urls need not be repeatedly archived. In these cases, taking
time to archive urls already in the IA just wastes the 15/minute
rate-limited saves.

To fix this, we must check if urls are already archived. The
Wayback Availability JSON API allows this and does not appear to be
rate-limited. See: https://archive.org/help/wayback_api.php

At present, I only check if an archive exists. I think loopholes exist
such as if an archive attempt was made, but the archive failed. To
fix this, I could check the snapshot status.

[fix] Add main file for update